requirements.txt
# Web Framework
Flask==3.1.2
Flask-WTF==1.2.1              # CSRF protection
WTForms==3.1.2                # Form handling
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
click==8.3.0
itsdangerous==2.2.0
blinker==1.9.0

# Security - XSS Protection
bleach==6.2.0
webencodings==0.5.1

# Security - Authentication
Argon2-cffi==23.1.0         # Password hashing (OWASP recommended)
cryptography==41.0.7        # Encryption for TOTP secrets
Flask-ReCaptcha==0.4.2     # Google reCAPTCHA integration

# Two-Factor Authentication
pyotp==2.9.0                # TOTP implementation
qrcode[pil]==7.4.2          # QR code generation
Pillow>=10.1.0              # Image processing for QR codes

# OAuth2
authlib==1.3.0              # OAuth2 provider/client

# Utilities
python-dotenv==1.0.0        # Environment variables
requests==2.31.0            # HTTP client (password breach checking)

# Terminal colors (existing)
colorama==0.4.6
